v0.4.0 - match data storage
Added system to find the player's matches and store the links to them, along with the player's corresponding rating, in a hash map.

Lowered minimum maps filter on players page to 50 from default 388 - this was useful for testing, as I could use players with far less data. In the future the program will likely start at 500 and iteratively lower the filter as the player is failed to be found.

Small readability improvements.